requirements.txt
attrs==23.1.0
black==23.7.0
boto3==1.28.18
wrapt==1.13.3
botocore==1.31.18
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
distlib==0.3.7
filelock==3.12.2
flake8==6.1.0
flake8-bugbear==20.1.4
flake8-print==3.1.4
greenlet==2.0.2
identify==2.5.26
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.3.3
pycodestyle==2.11.0
pyflakes==3.1.0
PyMuPDF==1.22.5
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
tomli==2.0.1
typing-extensions==4.7.1
urllib3==1.26.16
virtualenv==20.24.2
python-dotenv
